First and foremost: update to the new NTA API endpoint
Additionally, update some of the build infra components in order to build a
new release. This copies the bazel bits from the prometheus-dyson repo, which
are known to work (even though not the latest version of the python package for
bazel). The updates also allowed for the retirement of the aapipfix hack, which
is a nice bonus.