Support CANCELED and ADDED trips
Canceled trips will show in GetLive() with a new `canceled` field,
but will get filtered in GetUpcoming() (because it's not upcoming anymore).

Added trips work at least for PTV's MetroTrains, but this is a deprecated
field in GTFS Realtime and it's not clear there is standard usage. To show up
in live, an added trip needs to reference a stop we care about and a route
we know about. This required a bit of refactoring of the database API to
provide info on routes directly.